deploy/requirements.txt
# ☁️👜 aiClaudia - Python Dependencies

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.7

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2
